Feat(tutorial5) : Discovered the type conversio

+++ tutorial5/src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
     let x = 25 as i8;
-    let y = 10 as i8;
+    let y = 10 as i32;
 
-    let z = x / y;
-    println!("{}", z); // Finally gives 25.5
+    let z = x / (y as i8);
+    println!("{}", z);
 }
